requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-telegram-bot==0.6.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-yasg==1.21.5
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
requests==2.29.0
rfc3986==1.5.0
ruamel.yaml==0.17.22
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
